% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm.R
\name{calculate_icc_internal}
\alias{calculate_icc_internal}
\title{Calculate Intraclass Correlation}
\usage{
calculate_icc_internal(fit)
}
\arguments{
\item{fit}{Fitted lmer object}
}
\value{
Named vector of ICC values (one per grouping factor)
}
\description{
Calculate Intraclass Correlation
}
\keyword{internal}
