% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_mediators.R
\name{estimate_serial_paths}
\alias{estimate_serial_paths}
\title{Estimate Paths for Serial Mediation}
\usage{
estimate_serial_paths(data, x, mediators, y, covariates)
}
\description{
Estimate Paths for Serial Mediation
}
\keyword{internal}
