% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm.R
\name{extract_mlm_fit}
\alias{extract_mlm_fit}
\title{Extract MLM Model Fit Statistics}
\usage{
extract_mlm_fit(fit)
}
\arguments{
\item{fit}{Fitted lmer object}
}
\value{
Tibble of fit statistics
}
\description{
Extract MLM Model Fit Statistics
}
\keyword{internal}
