% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{extract_sem_fit}
\alias{extract_sem_fit}
\title{Extract SEM Fit Indices}
\usage{
extract_sem_fit(fit, measures)
}
\arguments{
\item{fit}{Fitted lavaan object}

\item{measures}{Character vector of fit measures to extract}
}
\value{
Tibble of fit indices with interpretation
}
\description{
Extract SEM Fit Indices
}
\keyword{internal}
