% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{extract_sem_indirect}
\alias{extract_sem_indirect}
\title{Extract Indirect Effects from SEM}
\usage{
extract_sem_indirect(fit)
}
\arguments{
\item{fit}{Fitted lavaan object}
}
\value{
Tibble of indirect effects with CIs, or NULL if none defined
}
\description{
Extract Indirect Effects from SEM
}
\keyword{internal}
