% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{extract_sem_r2}
\alias{extract_sem_r2}
\title{Extract R² for Endogenous Variables}
\usage{
extract_sem_r2(fit)
}
\arguments{
\item{fit}{Fitted lavaan object}
}
\value{
Tibble of R² values
}
\description{
Extract R² for Endogenous Variables
}
\keyword{internal}
