% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm.R
\name{generate_assumption_summary}
\alias{generate_assumption_summary}
\title{Generate MLM Assumption Summary}
\usage{
generate_assumption_summary(resid_normal, re_normality)
}
\arguments{
\item{resid_normal}{Logical, are residuals normal?}

\item{re_normality}{List of random effects normality tests}
}
\value{
Character string with summary
}
\description{
Generate MLM Assumption Summary
}
\keyword{internal}
