% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm.R
\name{generate_mlm_interpretation}
\alias{generate_mlm_interpretation}
\title{Generate MLM Interpretation Text}
\usage{
generate_mlm_interpretation(fixed, icc, n_groups)
}
\arguments{
\item{fixed}{Tibble of fixed effects}

\item{icc}{Named vector of ICC values}

\item{n_groups}{Named vector of group counts}
}
\value{
Character string with interpretation
}
\description{
Generate MLM Interpretation Text
}
\keyword{internal}
