% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlm.R
\name{generate_mlm_plots}
\alias{generate_mlm_plots}
\title{Generate MLM Diagnostic Plots}
\usage{
generate_mlm_plots(fit, resids, ranef_list)
}
\arguments{
\item{fit}{lmer fit object}

\item{resids}{Residuals}

\item{ranef_list}{List of random effects}
}
\value{
List of plots (or NULL if plotting not possible)
}
\description{
Generate MLM Diagnostic Plots
}
\keyword{internal}
