% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{generate_sem_interpretation}
\alias{generate_sem_interpretation}
\title{Generate SEM Interpretation Text}
\usage{
generate_sem_interpretation(fit_stats, params, indirect)
}
\arguments{
\item{fit_stats}{Tibble of fit statistics}

\item{params}{Tibble of parameters}

\item{indirect}{Tibble of indirect effects (or NULL)}
}
\value{
Character string with interpretation
}
\description{
Generate SEM Interpretation Text
}
\keyword{internal}
