% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_analysis.R
\name{print.alpha_analysis}
\alias{print.alpha_analysis}
\title{Print Method for Alpha Analysis}
\usage{
\method{print}{alpha_analysis}(x, ...)
}
\arguments{
\item{x}{An alpha_analysis object}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for Alpha Analysis
}
