% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_reliability.R
\name{print.composite_reliability}
\alias{print.composite_reliability}
\title{Print Method for Composite Reliability}
\usage{
\method{print}{composite_reliability}(x, ...)
}
\arguments{
\item{x}{A composite_reliability object}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for Composite Reliability
}
