% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_analysis.R
\name{print.efa_results}
\alias{print.efa_results}
\title{Print Method for EFA Results}
\usage{
\method{print}{efa_results}(x, ...)
}
\arguments{
\item{x}{An efa_results object}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for EFA Results
}
