% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{print.imported_data}
\alias{print.imported_data}
\title{Print Method for Imported Data}
\usage{
\method{print}{imported_data}(x, ...)
}
\arguments{
\item{x}{An imported_data object}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for Imported Data
}
