% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_mediators.R
\name{print.parallel_mediation}
\alias{print.parallel_mediation}
\title{Print Method for Parallel Mediation}
\usage{
\method{print}{parallel_mediation}(x, ...)
}
\description{
Print Method for Parallel Mediation
}
