% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_text.R
\name{print.publication_block}
\alias{print.publication_block}
\title{Print Method for Publication Blocks}
\usage{
\method{print}{publication_block}(x, ...)
}
\arguments{
\item{x}{A publication_block object}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for Publication Blocks
}
