% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{print.sem_result}
\alias{print.sem_result}
\title{Print Method for SEM Results}
\usage{
\method{print}{sem_result}(x, ...)
}
\arguments{
\item{x}{Object of class \code{sem_result}}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for SEM Results
}
