% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_mediators.R
\name{print.serial_mediation}
\alias{print.serial_mediation}
\title{Print Method for Serial Mediation}
\usage{
\method{print}{serial_mediation}(x, ...)
}
\description{
Print Method for Serial Mediation
}
